security updates

+++ src/main/java/com/tinder/controller/ConfirmEmail.java
@@ -15,18 +15,20 @@ import com.tinder.model.pojo.UnconfirmedUser;
 @Controller
 @RequestMapping("/ConfirmEmail")
 public class ConfirmEmail {
-	
+
 	@Autowired
 	private UserDAO userDAO;
-	
+
 	@RequestMapping(method = RequestMethod.GET)
-	public String confirmEmail(HttpServletRequest req,@RequestParam("UUID")String uuid){
-		UnconfirmedUser user = userDAO.getUnconfirmedUser(uuid);
-		userDAO.deleteUnconfirmedUser(uuid);
-		userDAO.registerUserWithHashedPassword(user.getUsername(),
-				user.getPasswordHash(),user.getEmail(),user.isGenderIsMale(),
-				user.getAge(), user.getFullName());
-		req.setAttribute("error", "Please confirm your email!");
+	public String confirmEmail(HttpServletRequest req, @RequestParam("UUID") String uuid) {
+		if (uuid != null) {
+			UnconfirmedUser user = userDAO.getUnconfirmedUser(uuid);
+			if (user != null) {
+				userDAO.deleteUnconfirmedUser(uuid);
+				userDAO.registerUserWithHashedPassword(user.getUsername(), user.getPasswordHash(), user.getEmail(),
+						user.isGenderIsMale(), user.getAge(), user.getFullName());
+			}
+		}
 		return "forward:/login";
 	}
 }
